requirements.txt
gradio>=4.0.0
langgraph>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers>=4.36.0
torch>=2.1.0
huggingface-hub>=0.19.0
PyPDF2>=3.0.1 